Move veres one method files into its own folder.
