+++ lib/index.js
@@ -6,7 +6,7 @@
 const Injector = require('./Injector');
 const injector = new Injector();
 
-const VeresOne = require('./methods/veres-one.js');
+const VeresOne = require('./methods/veres-one/veres-one.js');
 
 module.exports = {
   injector,
