+++ lib/methods/veres-one/veres-one.config.js
