+++ lib/methods/veres-one/veres-one.js
